Create TransactionInputPage, refactor code to make App.js stateless
