--- src/components/pages/TransactionInputPage.js
@@ -1,5 +1,0 @@
-import react from 'react'
-import styled from 'styled-components/macro'
-import TransactionFormInput from './TransactionFormInput'
-import TransactionList from './TransactionList'
-import Transaction from './Transaction'
